+++ exp/services/recoverysigner/internal/serve/account_delete_test.go
@@ -201,7 +201,13 @@ func TestAccountDelete_authenticatedByIdentityAddress(t *testing.T) {
 			"role": "receiver"
 		}
 	],
-	"signer": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE"
+	"signer": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE",
+	"signers": [
+		{
+			"key": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE",
+			"added_at": "0001-01-01T00:00:00Z"
+		}
+	]
 }`
 
 	assert.JSONEq(t, wantBody, string(body))
@@ -263,7 +269,13 @@ func TestAccountDelete_authenticatedByAccountAddress(t *testing.T) {
 			"role": "receiver"
 		}
 	],
-	"signer": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE"
+	"signer": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE",
+	"signers": [
+		{
+			"key": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE",
+			"added_at": "0001-01-01T00:00:00Z"
+		}
+	]
 }`
 
 	assert.JSONEq(t, wantBody, string(body))
@@ -326,7 +338,13 @@ func TestAccountDelete_authenticatedByPhoneNumber(t *testing.T) {
 			"authenticated": true
 		}
 	],
-	"signer": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE"
+	"signer": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE",
+	"signers": [
+		{
+			"key": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE",
+			"added_at": "0001-01-01T00:00:00Z"
+		}
+	]
 }`
 
 	assert.JSONEq(t, wantBody, string(body))
@@ -389,7 +407,13 @@ func TestAccountDelete_authenticatedByEmail(t *testing.T) {
 			"authenticated": true
 		}
 	],
-	"signer": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE"
+	"signer": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE",
+	"signers": [
+		{
+			"key": "GCAPXRXSU7P6D353YGXMP6ROJIC744HO5OZCIWTXZQK2X757YU5KCHUE",
+			"added_at": "0001-01-01T00:00:00Z"
+		}
+	]
 }`
 
 	assert.JSONEq(t, wantBody, string(body))