+++ services/horizon/internal/actions_operation.go
@@ -5,119 +5,19 @@ import (
 	"strings"
 
 	"github.com/stellar/go/services/horizon/internal/actions"
-	"github.com/stellar/go/services/horizon/internal/db2"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/ledger"
 	"github.com/stellar/go/services/horizon/internal/render/problem"
-	"github.com/stellar/go/services/horizon/internal/render/sse"
 	"github.com/stellar/go/services/horizon/internal/resourceadapter"
 	"github.com/stellar/go/services/horizon/internal/toid"
 	"github.com/stellar/go/support/errors"
 	"github.com/stellar/go/support/render/hal"
 	supportProblem "github.com/stellar/go/support/render/problem"
-	"github.com/stellar/go/xdr"
 )
-
-// This file contains the actions:
-//
-// OperationIndexAction: pages of operations
-// OperationShowAction: single operation by id
-
-// Interface verifications
-var _ actions.JSONer = (*OperationIndexAction)(nil)
-var _ actions.EventStreamer = (*OperationIndexAction)(nil)
 
 const (
 	joinTransactions = "transactions"
 )
-
-// OperationIndexAction renders a page of operations resources, identified by
-// a normal page query and optionally filtered by an account, ledger, or
-// transaction.
-type OperationIndexAction struct {
-	Action
-	LedgerFilter        int32
-	AccountFilter       string
-	TransactionFilter   string
-	PagingParams        db2.PageQuery
-	OperationRecords    []history.Operation
-	TransactionRecords  []history.Transaction
-	Ledgers             *history.LedgerCache
-	Page                hal.Page
-	IncludeFailed       bool
-	IncludeTransactions bool
-	OnlyPayments        bool
-}
-
-// JSON is a method for actions.JSON
-func (action *OperationIndexAction) JSON() error {
-	action.Do(
-		action.EnsureHistoryFreshness,
-		action.loadParams,
-		action.ValidateCursorWithinHistory,
-		action.loadRecords,
-		action.loadLedgers,
-		action.loadPage,
-		func() { hal.Render(action.W, action.Page) },
-	)
-	return action.Err
-}
-
-// SSE is a method for actions.SSE
-func (action *OperationIndexAction) SSE(stream *sse.Stream) error {
-	action.Setup(
-		action.EnsureHistoryFreshness,
-		action.loadParams,
-		action.ValidateCursorWithinHistory,
-	)
-	action.Do(
-		action.loadRecords,
-		action.loadLedgers,
-		func() {
-			stream.SetLimit(int(action.PagingParams.Limit))
-			operationRecords := action.OperationRecords[stream.SentCount():]
-			var transactionRecords []history.Transaction
-			if action.IncludeTransactions {
-				transactionRecords = action.TransactionRecords[stream.SentCount():]
-			}
-			for i, operationRecord := range operationRecords {
-				ledger, found := action.Ledgers.Records[operationRecord.LedgerSequence()]
-				if !found {
-					action.Err = errors.New(fmt.Sprintf("could not find ledger data for sequence %d", operationRecord.LedgerSequence()))
-					return
-				}
-
-				var transactionRecord *history.Transaction
-				if action.IncludeTransactions {
-					transactionRecord = &transactionRecords[i]
-				}
-
-				transactionHash := action.TransactionFilter
-				if len(transactionHash) == 0 {
-					transactionHash = operationRecord.TransactionHash
-				}
-				res, err := resourceadapter.NewOperation(
-					action.R.Context(),
-					operationRecord,
-					transactionHash,
-					transactionRecord,
-					ledger,
-				)
-				if err != nil {
-					action.Err = err
-					return
-				}
-
-				stream.Send(sse.Event{
-					ID:   res.PagingToken(),
-					Data: res,
-				})
-			}
-		},
-	)
-
-	return action.Err
-}
 
 func parseJoinField(action *actions.Base) (map[string]bool, error) {
 	join := action.GetString("join")
@@ -137,55 +37,7 @@ func parseJoinField(action *actions.Base) (map[string]bool, error) {
 	return validJoins, nil
 }
 
-func (action *OperationIndexAction) loadParams() {
-	action.ValidateCursorAsDefault()
-	action.AccountFilter = action.GetAddress("account_id")
-	action.LedgerFilter = action.GetInt32("ledger_id")
-	var err error
-	action.TransactionFilter, err = actions.GetTransactionID(action.R, "tx_id")
-	if err != nil {
-		action.Err = err
-		return
-	}
-	action.PagingParams = action.GetPageQuery()
-	action.IncludeFailed = action.GetBool("include_failed")
-	parsed, err := parseJoinField(&action.Action.Base)
-	if err != nil {
-		action.Err = err
-		return
-	}
-	action.IncludeTransactions = parsed[joinTransactions]
-
-	filters, err := countNonEmpty(
-		action.AccountFilter,
-		action.LedgerFilter,
-		action.TransactionFilter,
-	)
-
-	if err != nil {
-		action.Err = errors.Wrap(err, "Error in countNonEmpty")
-		return
-	}
-
-	if filters > 1 {
-		action.Err = supportProblem.BadRequest
-		return
-	}
-
-	// Double check TransactionFilter as it's used to determine if failed txs should be returned
-	if action.TransactionFilter != "" && !isValidTransactionHash(action.TransactionFilter) {
-		action.Err = supportProblem.MakeInvalidFieldProblem("tx_id", errors.New("Invalid transaction hash"))
-		return
-	}
-
-	if action.IncludeFailed && !action.App.config.IngestFailedTransactions {
-		err := errors.New("`include_failed` parameter is unavailable when Horizon is not ingesting failed " +
-			"transactions. Set `INGEST_FAILED_TRANSACTIONS=true` to start ingesting them.")
-		action.Err = supportProblem.MakeInvalidFieldProblem("include_failed", err)
-		return
-	}
-}
-
+// This check has been moved to history/operation with unit test
 func validateTransactionForOperation(transaction history.Transaction, operation history.Operation) error {
 	if transaction.ID != operation.TransactionID {
 		return errors.Errorf(
@@ -217,120 +69,6 @@ func validateTransactionForOperation(transaction history.Transaction, operation 
 	}
 
 	return nil
-}
-
-func (action *OperationIndexAction) loadRecords() {
-	q := action.HistoryQ()
-	ops := q.Operations()
-
-	switch {
-	case action.AccountFilter != "":
-		ops.ForAccount(action.AccountFilter)
-	case action.LedgerFilter > 0:
-		ops.ForLedger(action.LedgerFilter)
-	case action.TransactionFilter != "":
-		ops.ForTransaction(action.TransactionFilter)
-	}
-
-	// When querying operations for transaction return both successful
-	// and failed operations. We assume that because user is querying
-	// this specific transactions, she knows it's status.
-	if action.TransactionFilter != "" || action.IncludeFailed {
-		ops.IncludeFailed()
-	}
-
-	if action.IncludeTransactions {
-		ops.IncludeTransactions()
-	}
-
-	if action.OnlyPayments {
-		ops.OnlyPayments()
-	}
-
-	action.OperationRecords, action.TransactionRecords, action.Err = ops.Page(action.PagingParams).Fetch()
-	if action.Err != nil {
-		return
-	}
-
-	if action.IncludeTransactions && len(action.TransactionRecords) != len(action.OperationRecords) {
-		action.Err = errors.New("number of transactions doesn't match number of operations")
-		return
-	}
-
-	for i, o := range action.OperationRecords {
-		if !action.IncludeFailed && action.TransactionFilter == "" {
-			if !o.TransactionSuccessful {
-				action.Err = errors.Errorf("Corrupted data! `include_failed=false` but returned transaction in /operations is failed: %s", o.TransactionHash)
-				return
-			}
-
-			var resultXDR xdr.TransactionResult
-			action.Err = xdr.SafeUnmarshalBase64(o.TxResult, &resultXDR)
-			if action.Err != nil {
-				return
-			}
-
-			if !resultXDR.Successful() {
-				action.Err = errors.Errorf("Corrupted data! `include_failed=false` but returned transaction /operations is failed: %s %s", o.TransactionHash, o.TxResult)
-				return
-			}
-		}
-		if action.IncludeTransactions {
-			transaction := action.TransactionRecords[i]
-			action.Err = validateTransactionForOperation(transaction, o)
-			if action.Err != nil {
-				return
-			}
-		}
-	}
-}
-
-// loadLedgers populates the ledger cache for this action
-func (action *OperationIndexAction) loadLedgers() {
-	action.Ledgers = &history.LedgerCache{}
-	for _, op := range action.OperationRecords {
-		action.Ledgers.Queue(op.LedgerSequence())
-	}
-	action.Err = action.Ledgers.Load(action.HistoryQ())
-}
-
-func (action *OperationIndexAction) loadPage() {
-	for i, operationRecord := range action.OperationRecords {
-		ledger, found := action.Ledgers.Records[operationRecord.LedgerSequence()]
-		if !found {
-			msg := fmt.Sprintf("could not find ledger data for sequence %d", operationRecord.LedgerSequence())
-			action.Err = errors.New(msg)
-			return
-		}
-
-		var transactionRecord *history.Transaction
-		if action.IncludeTransactions {
-			transactionRecord = &action.TransactionRecords[i]
-		}
-
-		var res hal.Pageable
-		transactionHash := action.TransactionFilter
-		if len(transactionHash) == 0 {
-			transactionHash = operationRecord.TransactionHash
-		}
-		res, action.Err = resourceadapter.NewOperation(
-			action.R.Context(),
-			operationRecord,
-			transactionHash,
-			transactionRecord,
-			ledger,
-		)
-		if action.Err != nil {
-			return
-		}
-		action.Page.Add(res)
-	}
-
-	action.Page.FullURL = action.FullURL()
-	action.Page.Limit = action.PagingParams.Limit
-	action.Page.Cursor = action.PagingParams.Cursor
-	action.Page.Order = action.PagingParams.Order
-	action.Page.PopulateLinks()
 }
 
 // Interface verification