+++ services/friendbot/CHANGELOG.md
@@ -6,6 +6,10 @@ file.  This project adheres to [Semantic Versioning](http://semver.org/).
 As this project is pre 1.0, breaking changes may happen for minor version
 bumps.  A breaking change will get clearly notified in this log.
 
+## Unreleased
+
+* Log User-Agent header in request logs.
+
 ## [v0.0.2] - 2019-11-20
 
 ### Changed